LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/return-value/traverseTo-cross-document-preventDefault-expected.txt



FAIL traverseTo() promise never settle when preventDefault()ing the navigate event (cross-document) assert_equals: expected 2 but got 1
